Add files via upload
There is such a group of people, they accept God’s commission and bear their missions, spreading the gospel of the kingdom and thereby having suffered the resistance and insult from the religious world as well as the CCP government’s arrest and persecution. They even run the risk of their lives. Under such hard and severe circumstances……